backend/admin_panel/pyproject.toml
[project]
name = "streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.48.0",
    "python-dotenv>=1.1.1",
]
